requirements.txt
tk
pillow
ucimlrepo
scikit-learn
matplotlib
